Part 11

+++ app/Http/Controllers/PostsController.php
@@ -10,6 +10,17 @@ use DB; //Adding Database library for mysql approach query
 class PostsController extends Controller
 {
     /**
+     * Create a new controller instance.
+     * Prevent Guest from Performing Actions in Posts
+     *
+     * @return void
+     */
+    public function __construct()
+    {
+        $this->middleware('auth', ['except'=>['index', 'show']]); //Exception added...
+    }
+
+    /**
      * Display a listing of the resource.
      *
      * @return \Illuminate\Http\Response
@@ -85,6 +96,12 @@ class PostsController extends Controller
     public function edit($id)
     {
         $post = Post::find($id);
+
+        // Check for Correct User
+        if(auth()->user()->id !== $post->user_id) {
+            return redirect('/posts')->with('error', 'Unauthorized  Page');  
+        }
+
         return view('posts.edit')->with('post', $post);  
     }
 
@@ -122,6 +139,12 @@ class PostsController extends Controller
     public function destroy($id)
     {
         $post = Post::Find($id);
+
+        // Check for Correct User
+        if(auth()->user()->id !== $post->user_id) {
+            return redirect('/posts')->with('error', 'Unauthorized  Page');  
+        }
+
         $post->delete();
         return redirect(url('/posts'))->with('success', 'Post Deleted ');
     }
